backend/requirements.txt
flask
mavsdk
pyserial